📝refactor code clean up.
