+++ services/dataset.services.ts
@@ -1,5 +1,5 @@
 import { API } from "@/api"
-import { DatasetResponse } from "@/types"
+import { DatasetResponse } from "@/components/organisms/DataSetInternal/types"
 
 export const getDataset = async (key: string) =>
   API.get<DatasetResponse>(`/${key}`)
