--- types/index.ts
@@ -1,2 +1,0 @@
-export * from "./navMenuItem.types"
-export * from "./dataset.types"
