--- types/navMenuItem.types.ts
@@ -1,6 +1,0 @@
-export type NavMenuItem = {
-  key: string
-  icon: React.FunctionComponentElement<{}>
-  label: string
-  link: string
-}
